pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alive-progress==3.3.0",
    "datasets==4.4.1",
    "huggingface-hub==0.36.0",
    "matplotlib==3.10.8",
    "numpy==2.3.5",
    "pandas==2.3.3",
    "peft==0.18.0",
    "protobuf",
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "python-docx>=0.8.11",
    "requests==2.32.5",
    "scipy==1.16.3",
    "sentencepiece",
    "seaborn==0.13.2",
    "sentence-transformers==5.1.2",
    "spacy==3.8.11",
    "torch==2.6.0+cu124",
    "transformers==4.57.3",
    "notebook>=7.5.4",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.hatch.build.targets.wheel]
packages = ["src/scitrans"]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[tool.deptry]
ignore_notebooks = true
extend_ignore_obsolete = [
    "pytest",
    "ipykernel",
    "notebook",
    "torchvision",
    "torchaudio"
]
